Move to package
Split out from #5.
