+++ setup.cfg
@@ -32,7 +32,7 @@ license_file = LICENSE
 [options]
 package_dir=
     =src
-py_modules = django_htmx
+packages = find:
 include_package_data = True
 install_requires =
     Django>=2.2
@@ -42,10 +42,9 @@ zip_safe = False
 [options.packages.find]
 where = src
 
-
 [coverage:run]
 branch = True
-include = */django_htmx.py
+include = */django_htmx/*.py
 
 [coverage:paths]
 source =
